Use commands to find exact floating IP address
Using these commands will make sure we grab a floating IP address that
is not already allocated to another user. This is done by making sure
the floating IP address we search for has a description that matches the
calling user's cluster name i.e. <username>-ocp and purpose for the IP
i.e. api or app.